# Optimization: Don't pass not-readonly `PackedFieldDescriptor` with `in` specifier