See explanation at https://devblogs.microsoft.com/premier-developer/the-in-modifier-and-the-readonly-structs-in-c/

<b>you should never pass a non-readonly struct as in parameter</b>
||                 Method |      Mean |     Error |    StdDev | 
|----------------------- |----------:|----------:|----------:| 
|AggregatePassedByValue |  71.24 us | 0.3150 us | 0.2278 us | 
|    AggregatePassedByIn | 124.02 us | 3.2885 us | 9.6963 us |

And making this struct `readonly` requires more code rewriting